Major things Updated

+++ src/Components/Postbox/PostBox.jsx
@@ -10,6 +10,8 @@ import { MdFullscreen } from "react-icons/md";
 import { HiOutlineDotsCircleHorizontal } from "react-icons/hi";
 import '../Home/Home.css';
 import YouTubeEmbed from '../YouTubeEmbed/YouTubeEmbed';
+import Swal from 'sweetalert2'
+import './PostBox.css'
 
 const PostBox = ({ post, onDelete }) => {
 
@@ -18,6 +20,7 @@ const PostBox = ({ post, onDelete }) => {
     const [reactInfo, setReactInfo] = useState([])
     const [liveCommentCount, setLiveCommentCount] = useState(post.commentOnPost.length);
     const id = post?.uploaderUid;
+    const [thisPost, setThisPost] = useState(post);
 
     useEffect(() => {
         fetch(`https://cse-p-diu-server.vercel.app/users/uid/${id}`)
@@ -34,7 +37,18 @@ const PostBox = ({ post, onDelete }) => {
     function copyLink(link) {
         navigator.clipboard.writeText(link)
             .then(() => {
-                alert('Link copied to clipboard!');
+                Swal.fire({
+                    position: "top-end",
+                    title: "Link Copied",
+                    showConfirmButton: false,
+                    color: '#ffffff',
+                    background: '#00ff0090',
+                    timer: 1000,
+                    toast: true,
+                    backdrop: false,
+                    timerProgressBar: true,
+                    width: 'auto',
+                });
             })
             .catch(err => {
                 console.error('Failed to copy link: ', err);
@@ -58,21 +72,79 @@ const PostBox = ({ post, onDelete }) => {
     };
 
     const handleDeletePost = () => {
-        fetch(`https://cse-p-diu-server.vercel.app/posts/${post?._id}`, {
-            method: 'DELETE',
+
+        Swal.fire({
+            title: "Are you sure?",
+            text: "You won't be able to revert this!",
+            icon: "warning",
+            showCancelButton: true,
+            color: "#ffffff",
+            background: '#1d232a60',
+            confirmButtonColor: "#3085d6",
+            cancelButtonColor: "#d33",
+            confirmButtonText: "Yes, delete it!"
+        }).then((result) => {
+            if (result.isConfirmed) {
+                fetch(`https://cse-p-diu-server.vercel.app/posts/${post?._id}`, {
+                    method: 'DELETE',
+                    headers: {
+                        'content-type': 'application/json',
+                    },
+                })
+                    .then(res => res.json())
+                    .then(data => {
+                        // console.log(data);
+                        onDelete(post._id);
+                    })
+                    .catch(error => {
+                        // console.log(error);
+                    })
+                Swal.fire({
+                    title: "Deleted!",
+                    text: "Your file has been deleted.",
+                    icon: "success",
+                    background: '#1d232a60',
+                    color: "#ffffff",
+                });
+            }
+        });
+    }
+
+
+
+
+    const handleEditPost = e => {
+        e.preventDefault();
+        const form = e.target;
+        const postId = post._id;
+        const updatedTitle = form.updatedTitle.value;
+        const updatedDetails = form.updatedDetails.value;
+        const updatedPost = { postId, updatedTitle, updatedDetails };
+
+        fetch(`https://cse-p-diu-server.vercel.app/posts/${postId}`, {
+            method: 'PUT',
             headers: {
                 'content-type': 'application/json',
             },
+            body: JSON.stringify(updatedPost)
         })
             .then(res => res.json())
             .then(data => {
+                document.getElementById(`postEditingModal-${thisPost._id}`).close();
                 // console.log(data);
-                onDelete(post._id);
+                if (data.modifiedCount >= 1) {
+                    setThisPost(prevPost => ({
+                        ...prevPost,
+                        postTitle: updatedTitle,
+                        postDetails: updatedDetails,
+                    }));
+                }
             })
             .catch(error => {
-                // console.log(error);
-            })
+                console.error('Error updating post:', error);
+            });
     }
+
 
     const handleReacting = () => {
         const hasUserReacted = reactInfo.some(react => react.userUid === user.uid);
@@ -127,14 +199,14 @@ const PostBox = ({ post, onDelete }) => {
         setLiveCommentCount(newCount);
     };
 
-    
+
 
 
     return (
         <div className="w-full">
             <div className="flex flex-col w-full rounded-2xl shadow-lg p-3 sm:p-[2px] cursor-pointer transition duration-150 mt-2 ease-in-out ">
                 <div className="flex justify-between p-5 font-bold border-l border-t border-r border-[#414141] rounded-t-lg text-center">
-                    <p className=" ">{post?.postTitle}</p>
+                    <p className=" ">{thisPost?.postTitle}</p>
                     <div className="flex flex-row">
                         {user?.uid === id ?
                             (<div className="dropdown dropdown-end">
@@ -146,14 +218,52 @@ const PostBox = ({ post, onDelete }) => {
                                 <ul
                                     tabIndex={0}
                                     className="menu menu-sm dropdown-content bg-base-100 rounded-box z-[1] mt-3 w-52 p-2 shadow">
-                                    <button className="rounded hover:bg-[#ffffff06] mt-1 transition duration-150 ease-in-out">Edit Post</button>
+                                    <button onClick={() => document.getElementById(`postEditingModal-${thisPost._id}`).showModal()} className="rounded hover:bg-[#ffffff06] mt-1 transition duration-150 ease-in-out">Edit Post</button>
                                     <button onClick={handleDeletePost} className="rounded hover:bg-[#ffffff06] mt-1 transition duration-150 ease-in-out">Delete</button>
                                 </ul>
                             </div>)
                             : ''}
-                        <div className="text-2xl hover:scale-110"><Link to={`/post/${post._id}`} ><MdFullscreen /></Link></div>
+                        <div className="text-2xl hover:scale-110"><Link to={`/post/${thisPost._id}`} ><MdFullscreen /></Link></div>
                     </div>
                 </div>
+
+
+
+
+                <dialog id={`postEditingModal-${thisPost._id}`} className="modal">
+                    <div className="modal-box">
+                        <form onSubmit={handleEditPost}>
+                            <form method="dialog">
+                                {/* if there is a button in form, it will close the modal */}
+                                <button className="btn btn-sm btn-circle btn-ghost absolute right-2 top-2"><AiOutlineClose /></button>
+                            </form>
+                            <h3 className="font-bold text-lg">Edit Post</h3>
+
+                            <label className="form-control w-full">
+                                <div className="label">
+                                    <span className="label-text">Post Title</span>
+                                </div>
+                                <input defaultValue={thisPost?.postTitle} type="text" name="updatedTitle" placeholder="Post Title" className="input input-bordered w-full" required />
+                            </label>
+
+                            <label className="form-control w-full">
+                                <div className="label">
+                                    <span className="label-text">Post Details</span>
+                                </div>
+                                <textarea rows='6' defaultValue={thisPost?.postDetails} name="updatedDetails" className="textarea textarea-bordered" placeholder="Post Details........."></textarea>
+                            </label>
+                            <div className="text-center">
+                                <input className="btn mt-6" type="submit" value="Update" />
+                            </div>
+                        </form>
+
+
+
+                    </div>
+                </dialog>
+
+
+
 
                 <img src={post?.imageUrl} alt="" />
 
@@ -217,7 +327,7 @@ const PostBox = ({ post, onDelete }) => {
                             )
                                 :
                                 <div>
-                                    <h2>Please Login to Comment/React</h2>
+                                    <h2>Please <Link className='text-[#4444ff] font-bold' to='/login'>Login</Link> to Comment/React</h2>
                                     <div>
                                         <button onClick={() => copyLink(`https://cse-p-diu.web.app/post/${post._id}`)} className="text-3xl mx-4 hover:scale-110">
                                             <IoCopyOutline />
